Small Fixes
I have fixed a bug when trying to get images from Plex that is on a
different server.
Also fixed the drive spaces on macs, and disabled memory for mac until I
work out a better solution.

Added a new option in the config.js file
runningMode, normal or debug.